Implement user authentication and workout management features; add user_id to workouts, requireAuth middleware, and enhance routing with protected routes.

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,12 +1,21 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuth } from "../hooks/useAuth";
+
 
 const WorkoutDetails = ({workout}) => {
     const {dispatch} = useWorkoutsContext();
+    const {user} = useAuth();
 
     const handleClick = async () => {
+        if(!user) {
+            return
+        }
         const response = await fetch ('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
 
         const json = await response.json();
